move lesson copy service into lesson module
