+++ apps/server/src/modules/copy-helper/index.ts
@@ -1,4 +1,5 @@
 export * from './copy-helper.module';
 export * from './copy-helper.service';
+export * from './copy.mapper';
 export * from './copy.response';
 export * from './copy.types';
