+++ apps/server/src/modules/files-storage-client/files-storage-client.module.ts
@@ -7,7 +7,6 @@ import { FilesStorageProducer } from './service/files-storage.producer';
 
 @Module({
 	imports: [LoggerModule, CopyHelperModule],
-	controllers: [],
 	providers: [FilesStorageClientAdapterService, CopyFilesService, FilesStorageProducer],
 	exports: [FilesStorageClientAdapterService, CopyFilesService],
 })
