+++ apps/server/src/modules/learnroom/controller/rooms.controller.ts
@@ -4,11 +4,11 @@ import { RequestTimeout } from '@shared/common';
 import { ICurrentUser } from '@shared/domain';
 import { Authenticate, CurrentUser } from '@src/modules/authentication/decorator/auth.decorator';
 import { CopyApiResponse } from '@src/modules/copy-helper';
-import { LessonCopyUC } from '@src/modules/learnroom/uc/lesson-copy.uc';
+import { CopyMapper } from '@src/modules/copy-helper/copy.mapper';
 import { serverConfig } from '@src/modules/server/server.config';
-import { CopyMapper } from '../mapper/copy.mapper';
 import { RoomBoardResponseMapper } from '../mapper/room-board-response.mapper';
 import { CourseCopyUC } from '../uc/course-copy.uc';
+import { LessonCopyUC } from '../uc/lesson-copy.uc';
 import { RoomsUc } from '../uc/rooms.uc';
 import {
 	BoardResponse,
