+++ apps/server/src/modules/learnroom/service/index.ts
@@ -1,3 +1,2 @@
 export * from './board-copy.service';
 export * from './course-copy.service';
-export * from './lesson-copy.service';
