+++ apps/server/src/modules/lesson/index.ts
@@ -1,3 +1,5 @@
-export * from './lesson.module';
 export * from './lesson-api.module';
+export * from './lesson.module';
+export * from './types/lesson-copy-parent.params';
+export * from './types/lesson-copy.params';
 export * from './uc';
