+++ apps/server/src/modules/lesson/service/etherpad.service.ts
