+++ apps/server/src/modules/lesson/service/index.ts
@@ -1,1 +1,4 @@
+export * from './etherpad.service';
+export * from './lesson-copy.service';
 export * from './lesson.service';
+export * from './nexboard.service';
