+++ apps/server/src/modules/lesson/types/index.ts
@@ -1,0 +1,2 @@
+export * from './lesson-copy-parent.params';
+export * from './lesson-copy.params';
