+++ apps/server/src/modules/task/controller/task.controller.ts
@@ -5,8 +5,8 @@ import { PaginationParams } from '@shared/controller/';
 import { ICurrentUser } from '@shared/domain';
 import { Authenticate, CurrentUser } from '@src/modules/authentication/decorator/auth.decorator';
 // TODO  @src/modules/learnroom/* must be replaced
+import { CopyMapper } from '@src/modules/copy-helper';
 import { CopyApiResponse } from '@src/modules/copy-helper/copy.response';
-import { CopyMapper } from '@src/modules/learnroom/mapper/copy.mapper';
 import { serverConfig } from '@src/modules/server/server.config';
 import { TaskMapper } from '../mapper';
 import { TaskCopyUC, TaskUC } from '../uc';
