+++ apps/server/src/modules/task/types/task-copy-parent.params.ts
@@ -1,0 +1,8 @@
+import { EntityId } from '@shared/domain';
+
+// todo: it look like it is required not optional
+export type TaskCopyParentParams = {
+	courseId?: EntityId;
+	lessonId?: EntityId;
+	userId: string;
+};
