+++ apps/server/src/shared/domain/service/index.ts
@@ -1,3 +1,1 @@
-export * from './etherpad.service';
-export * from './nexboard.service';
 export * from './permission.service';
